Hotifx/Correção de bugs

+++ item.c
@@ -3,12 +3,12 @@
 #include "item.h"
 
 struct item_{
-    float peso;
-    float valor;
+    int peso;
+    int valor;
     int chave;
 };
 
-ITEM *ItemCriar(float peso, float valor, int chave){
+ITEM *ItemCriar(int peso, int valor, int chave){
     //Cria o item
     ITEM *novoItem = (ITEM *)calloc(1, sizeof(ITEM));
     if(novoItem == NULL)
@@ -21,19 +21,19 @@ ITEM *ItemCriar(float peso, float valor, int chave){
     return novoItem;
 }
 
-float ItemGetValor(ITEM *item){
+int ItemGetValor(ITEM *item){
     if(item != NULL)
         return item->valor;
     return -1;
 }
 
-float ItemGetPeso(ITEM *item){
+int ItemGetPeso(ITEM *item){
     if(item != NULL)
         return item->peso;
     return -1;
 }
 
-float ItemGetRazao(ITEM *item){
+int ItemGetRazao(ITEM *item){
     if(item != NULL)
         return (item->valor / item->peso);
     return -1;
