+++ mochila.h
@@ -5,10 +5,10 @@
 
     typedef struct mochila_ MOCHILA;
 
-    MOCHILA *MochilaCriar(float pesoMaximo);
+    MOCHILA *MochilaCriar(int pesoMaximo);
     ITEM *MochilaRemoverItem(MOCHILA *mochila, int chave);
-    float MochilaGetValor(MOCHILA *mochila);
-    float MochilaGetPeso(MOCHILA *mochila);
+    int MochilaGetValor(MOCHILA *mochila);
+    int MochilaGetPeso(MOCHILA *mochila);
     int MochilaGetQuantidadeItens(MOCHILA *mochila);
     bool MochilaAdicionarItem(MOCHILA *mochila, ITEM *item);
     bool MochilaCabe(MOCHILA *mochila, ITEM *item);
